problem6 create duplicatedChecker
